色々変更 Double → Decimalに変更 Modelについては保存形式はStringにしてget set で Decimalを扱うように変更 アプリ名変更 アイコン追加

+++ SumCalculator/Model/CalcNote.swift
@@ -21,9 +21,9 @@ class CalcNote: Object {
             editedAt = Date()
         }
     }
-    var total: Double {
+    var total: Decimal {
         get {
-            var sum = Double()
+            var sum = Decimal()
             for calcTable in calcTables {
                 sum += calcTable.subtotal
             }
